# lib/api-zod/src/generated/types/intelligenceV2Run.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { IntelligenceV2CommercialRole } from './intelligenceV2CommercialRole';
import type { IntelligenceV2Evidence } from './intelligenceV2Evidence';
import type { IntelligenceV2RunCost } from './intelligenceV2RunCost';
import type { IntelligenceV2RunDeterministicOverridesItem } from './intelligenceV2RunDeterministicOverridesItem';
import type { IntelligenceV2RunFingerprints } from './intelligenceV2RunFingerprints';
import type { IntelligenceV2RunIdentity } from './intelligenceV2RunIdentity';
import type { IntelligenceV2RunIntelligenceVersion } from './intelligenceV2RunIntelligenceVersion';
import type { IntelligenceV2RunPrimaryBusiness } from './intelligenceV2RunPrimaryBusiness';
import type { IntelligenceV2RunResolutionType } from './intelligenceV2RunResolutionType';
import type { IntelligenceV2RunVersions } from './intelligenceV2RunVersions';
import type { IntelligenceV2Who } from './intelligenceV2Who';

export interface IntelligenceV2Run {
  intelligenceVersion: IntelligenceV2RunIntelligenceVersion;
  projectId: string;
  projectCompanyId: string;
  companyId: string;
  companyName: string;
  /** @nullable */
  domain: string | null;
  createdAt: string;
  identity: IntelligenceV2RunIdentity;
  /** @nullable */
  primaryBusiness: IntelligenceV2RunPrimaryBusiness;
  commercialRole: IntelligenceV2CommercialRole;
  who: IntelligenceV2Who;
  /**
     * @minimum 0
     * @maximum 1
     */
  assessmentConfidence: number;
  resolutionType: IntelligenceV2RunResolutionType;
  deterministicOverrides: IntelligenceV2RunDeterministicOverridesItem[];
  unknownFacts: string[];
  evidence: IntelligenceV2Evidence[];
  /** Real currency. Present for internal admins only. */
  cost?: IntelligenceV2RunCost;
  versions: IntelligenceV2RunVersions;
  fingerprints: IntelligenceV2RunFingerprints;
}