# lib/api-zod/src/generated/types/projectChange.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { ChangeEvidenceRef } from './changeEvidenceRef';
import type { ChangeScore } from './changeScore';
import type { ChangeVerdict } from './changeVerdict';
import type { ProjectChangeTrigger } from './projectChangeTrigger';

export interface ProjectChange {
  id: string;
  projectCompanyId: string;
  companyId: string;
  companyName: string;
  domain: string | null;
  trigger: ProjectChangeTrigger;
  observedAt: Date;
  hasChanges: boolean;
  profileChanged: boolean;
  verdictChanged: boolean;
  scoreChanged: boolean;
  evidenceAdded: ChangeEvidenceRef[];
  evidenceRemoved: ChangeEvidenceRef[];
  evidenceChanged: ChangeEvidenceRef[];
  verdictBefore: ChangeVerdict | null;
  verdictAfter: ChangeVerdict;
  scoreBefore: ChangeScore | null;
  scoreAfter: ChangeScore | null;
  factsAdded: number;
  signalsCreated: number;
  modelCalls: number;
  /** Real currency. Present for internal admins only. */
  costTotal?: number;
}